Move dummy sim overlay to tests module
